Dbm 1554 - New Navigator (#43)
* Created hooks to get data from entity

* Entity selection happens in the 3rd panel now

> This change is dependent of a change made with branch DBM-1557 in ipa-dt

The entities were previously selected (appeared on the bottom panel and highlighted in the viewer), when filtered in the second panel. It seemed more intuitive that the first two panels only searched and filtered the entities that are displayed in the third panel, and isolated the elements in the model. The third panel is used to select entities, which means hilighting them in the model, and displaying them in the bottom panel.
Right now the bottom panel can only display the last entity selected, the ability to switch between selected entities in the bottom panel will be implemented later.

* EntityListView had some issue when used uncontrolled

* We can clear search and filters on the new navigator

* Added tools to ease fetching and displaying entity data

* The new navigator view can display extended data

- Created EntityDataGroupContainer to display entity data while being style agnostic
-  Modified withEntitySearch to load extended data when loading multiple entities
- Created SimpleTabbedTable component to display groups in tabs.

* Better tooltips

* EntityModal can now handle the delete action

* The delete action can now be performed on multiple entities even if unchanged

* Fixed entity tree loading

* Fixed : Entity action called changeEntity twice

* Fixed the upload action and switching to spaces when an asset is selected

* The new refreshing treesearch is not really working yet

This commit is to save my progress before trying out something else

* Fixed TreeSearch reloading when entities change

* EntityListView: exposed sortEntities custom hook / added onSortChange callback (#42)

Co-authored-by: dominika <dominika.oles@invicara.com>

Co-authored-by: doles <oles.dominika@gmail.com>
Co-authored-by: dominika <dominika.oles@invicara.com>