[DBM-1892] Fixed the eneity selection from model
The second entity selected was not displayed in the bottom panel, the first entity data was displayed instead

We were not projecting the id correctly in the setIncludesBy function params, the _id did not exist, I used modelData.id instead.

I hope it doesn't break anything.

I tested on a dsl project and a jsproject to make sure it was not the script that gave us the id in a different place on a js project.